added index and main page
