+++ client/src/Login/Login.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import Footer from '../Common/Footer/Footer';
+import Header from '../Common/Header/Header';
+import LoginPage from './LoginPage/LoginPage';
+function Login() {
+  return (
+    <>
+        <Header/>
+        <LoginPage/>
+        <Footer/>
+    </>
+  )
+}
+
+export default Login
