+++ client/src/Login/LoginPage/LoginPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function LoginPage() {
+  return (
+    <div>LoginPage</div>
+  )
+}
+
+export default LoginPage
