+++ client/src/Main/Main.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import MainPage from './MainPage/MainPage'
+function Main() {
+  return (
+    <>
+        <MainPage/>
+    </>
+  )
+}
+
+export default Main
