+++ client/src/Messaging/LeftPane/Contacts/Contact/MessageCount/MessageCount.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function MessageCount() {
+  return (
+    <div>MessageCount</div>
+  )
+}
+
+export default MessageCount
